requirements.txt
asgiref @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_edg6nzjrx0/croots/recipe/asgiref_1661287549316/work
certifi @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_477u68wvzm/croot/certifi_1671487773341/work/certifi
crispy-bootstrap5==0.7
Django @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_02mqkg06ol/croots/recipe/django_1661284244145/work
django-crispy-forms==2.0
djangorestframework==3.14.0
dnspython==2.3.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
mysqlclient==2.1.1
psycopg2 @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_c53vfly7f7/croot/psycopg2_1666817457628/work
pymongo==4.4.0
pytz @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_ddzpsmm2_f/croot/pytz_1671697430473/work
sqlparse==0.2.4
typing_extensions @ file:///private/var/folders/sy/f16zz6x50xz3113nwtb9bvq00000gp/T/abs_4b7xacf029/croot/typing_extensions_1669923792404/work
